Refactor /util/data into /util/offenses
The only remaining thing in the data util was a mapping of crime types.
I moved that to the offenses util and removed data.js.

Added propType declarations since I was in the file.
